V2 - 3 bytes size for signatures (#153)
* Add readUint24 utility function

* Use uint24 for chained signature size

* Use 3 bytes for dynamic signature

* Use 3 bytes for signature branchs

+++ foundry_test/modules/commons/submodules/auth/SequenceBaseSig.t.sol
@@ -152,7 +152,7 @@ contract SequenceBaseSigTest is AdvTest {
         if (op == 1) {
           signature = abi.encodePacked(signature, FLAG_SIGNATURE, randomWeight, sigpart);
         } else {
-          signature = abi.encodePacked(signature, FLAG_DYNAMIC_SIGNATURE, randomWeight, addr, uint16(sigpart.length), sigpart);
+          signature = abi.encodePacked(signature, FLAG_DYNAMIC_SIGNATURE, randomWeight, addr, uint24(sigpart.length), sigpart);
         }
       }
 
@@ -175,7 +175,7 @@ contract SequenceBaseSigTest is AdvTest {
 
     for (uint256 i = 0; i < size; i++) {
       if (i != 0) {
-        signature = abi.encodePacked(FLAG_BRANCH, uint16(signature.length), signature);
+        signature = abi.encodePacked(FLAG_BRANCH, uint24(signature.length), signature);
       }
 
       _pks[i] = boundPk(_pks[i]);
@@ -196,7 +196,7 @@ contract SequenceBaseSigTest is AdvTest {
         if (op == 1) {
           signature = abi.encodePacked(FLAG_SIGNATURE, randomWeight, sigpart, signature);
         } else {
-          signature = abi.encodePacked(FLAG_DYNAMIC_SIGNATURE, randomWeight, addr, uint16(sigpart.length), sigpart, signature);
+          signature = abi.encodePacked(FLAG_DYNAMIC_SIGNATURE, randomWeight, addr, uint24(sigpart.length), sigpart, signature);
         }
       }
