+++ test/utils/sequence.ts
@@ -475,7 +475,7 @@ export class SignatureConstructor {
         case SignaturePartType.Dynamic:
           const signature = ethers.utils.arrayify(member.value ?? [])
           result = ethers.utils.solidityPack(
-            ['bytes', 'uint8', 'uint8', 'address', 'uint16', 'bytes'],
+            ['bytes', 'uint8', 'uint8', 'address', 'uint24', 'bytes'],
             [result, SignaturePartType.Dynamic, member.weight, member.address, signature.length, signature]
           )
           break;
@@ -490,7 +490,7 @@ export class SignatureConstructor {
         case SignaturePartType.Branch:
           const branch = ethers.utils.arrayify(member.value ?? [])
           result = ethers.utils.solidityPack(
-            ['bytes', 'uint8', 'uint16', 'bytes'],
+            ['bytes', 'uint8', 'uint24', 'bytes'],
             [result, SignaturePartType.Branch, branch.length, branch]
           )
           break
